Create Stack syntax learning.java
Declare an Empty  .
Take Single Integer  as input.
For next  Lines format 
case .  the  of the  in a separate line.
case .  an element from the stack. If the stack is empty then print  in a separate line.
case .  Integer  to the  .
case 4.  an element at the  of the . If stack is empty print  in a seperate line.
Input Format

Single Integer .
 Lines of format .
Constraints

1 <= T <= 10^4
0 <= x <= 10^8
Output Format

According to case Print the information explained, if its not possible then -1.
Sample Input 0

10
3 1
3 2
4
4
2
4
3 4
2
4
1
Sample Output 0

2
2
1
1
1
Explanation 0

Based on instructions given, desired Output is obtained.