Create Next greater element on left 1.java
Given an array , print the Next Greater Element (NGE) for every element. The Next greater Element for an element  is the first greater element on the left side of  in the array. Elements for which no greater element exist, consider the next greater element as .

Input Format

First line contains an integer  representing the size of the array.
Next  lines contains elements of the array.
Constraints

1 <= N <= 1000
-1000 <= arr[i] <= 1000
Output Format

Print the answer array seperated by a single space
Sample Input 0

5
3 4 3 2 1
Sample Output 0

-1 -1 4 3 2
Explanation 0

No element on left of  = , which is greater so  = .
No element on left of  = , which is greater so  = .
 is nearest element to the left of  = , which is greater so  = .
 is nearest element to the left of  = , which is greater so  = .
 is nearest element to the left of  = , which is greater so  = .